Add files via upload
This is a simple weatherApp i decided to create with the basic knowledge gained from Hackerboost Web develoment bootcamp. This App is crested with Html,CSS and Javascript. the code wasnt a problem at all but the weather API keys were stressing out but the joy i got after debugging it unimaginable . By why this app at all? i decided to come out with this app because at times when i goes to a village and i want to track the weather conditions there has always been a challenge and coming out with this app will give an optimal solution to this